react-router-redux change error dispatch in actions

+++ app/components/Dashboard.jsx
@@ -1,14 +1,21 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import * as actions from 'actions';
+import {push} from 'react-router-redux';
+import Item from 'Item';
 
 class DashBoard extends Component {
     constructor(props) {
         super(props);
     }
     componentDidMount() {
-        this.props.dispatch(actions.deleteItems());
-        this.props.dispatch(actions.fetchItems());
+        if (this.props.auth.token) {
+            this.props.dispatch(actions.deleteItems());
+            this.props.dispatch(actions.fetchItems());
+        } else {
+            this.props.dispatch(push('/'));
+            this.props.dispatch(actions.setError('You must log in.'))
+        }
     }
     render() {
         let { items } = this.props;
@@ -18,22 +25,7 @@ class DashBoard extends Component {
                 return
             }
             return items.map((item) => {
-                let { action, author } = item.logs[item.logs.length - 1];
-                return (
-                    <div className="dashboard-item" key={key++}>
-                        <h3>{item.name}</h3>
-                        <div className="info-container">
-                            <span>Amount: {item.number}</span>
-                            <span>{item.state}</span>
-                        </div>
-                        <span className="created">{`${action} by ${author}`}</span>
-                        <span className="last-log">{`Last log: ${item.logs[0].action} by ${item.logs[0].author}`}</span>
-                        <div className="buttons">
-                            {/*<Link to={`/items/${item.id}/edit`}>Edit</Link>
-                            <Link to='/'>Delete</Link>*/}
-                        </div>
-                    </div>
-                );
+                return (<Item key={key++} item={item}/>)
             })
         }
             return (
